# add benchmark for Settings.toInternal + optimizations