--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,10 @@
+module github.com/segmentio/go-hll
+
+go 1.12
+
+require (
+	github.com/davecgh/go-spew v1.1.0
+	github.com/pkg/errors v0.8.0
+	github.com/pmezard/go-difflib v1.0.0
+	github.com/stretchr/testify v1.2.2
+)