Adds a benchmark and cleaned up a tiny bit of code. Clean code runs fast apparently 😛
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkSettingsToInternal-12     64.0          31.9          -50.16%

benchmark                          old allocs     new allocs     delta
BenchmarkSettingsToInternal-12     0              0              +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkSettingsToInternal-12     0             0             +0.00%
```